Split abenc_test into multiple files, one for each scheme
